// projects/mmb-web-app/src/app/profile/types/helpers/experience.model.ts
import { DurationItem } from './duration-item.model';

export class Experience {
    start_date: string;
    end_date: string;
    company: string;
    organisation_custom: string;
    organisation: string;
    title: string;
    logo: string;
    location: string;
    summary: string;
    uuid: string;
}