// projects/mmb-web-app/src/app/profile/types/helpers/education.model.ts
import { DurationItem } from './duration-item.model';

export class Education {
    start_date: string;
    end_date: string;
    field_of_study: string;
    institute: string;
    degree: string;
    uuid: string;
    logo: string;
    courseTitle: string;
    location: string;
    summary: string;
    title: string;
    organisation_custom: string;
    organisation: string;
    experience_type: number;
    notes: string;
}